+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - **[NEW]** Add `Card` util component, add `QrCard` component and use `Card` in `TripCard`
+- **[FIX]** Fix `MainContent` height when inside a `Modal` with close button
 - [...]
 
 # v23.2.0 (20/03/2020)